修改注释

+++ database/migrations/2020_02_28_171310_add_favorite_count_to_topics_table.php
@@ -15,7 +15,7 @@ class AddFavoriteCountToTopicsTable extends Migration
     {
         Schema::table('topics', function (Blueprint $table) {
               // 新增字段  收藏数量  after将此字段放置在view_count"之后" (MySQL)
-            $table->bigInteger('favorite_count')->after('view_count')->default(0)->comment('收藏数量');
+            $table->Integer('favorite_count')->after('view_count')->default(0)->comment('收藏数量');
         });
     }
